# Improve Int parsing performance